requirements.txt
streamlit
torch
timm
numpy
Pillow
albumentations
matplotlib
seaborn
requests